Renamed lib/cli/lookupFiles.js to lib/cli/lookup-files.js
