+++ lib/cli/collect-files.js
@@ -5,7 +5,7 @@ const ansi = require('ansi-colors');
 const debug = require('debug')('mocha:cli:run:helpers');
 const minimatch = require('minimatch');
 const {NO_FILES_MATCH_PATTERN} = require('../errors').constants;
-const lookupFiles = require('./lookupFiles');
+const lookupFiles = require('./lookup-files');
 
 /**
  * Exports a function that collects test files from CLI parameters.
