+++ lib/cli/lookup-files.js
